Add 2.0.0 version to client tests, clean up code, switch to use SuiteEndpoint functionality instead of record_response_route
