+++ lib/carin_for_blue_button_test_kit.rb
@@ -2,4 +2,4 @@ require_relative 'inferno_requirements_tools/ext/inferno_core/runnable'
 require_relative 'carin_for_blue_button_test_kit/generated/v1.1.0/c4bb_test_suite'
 require_relative 'carin_for_blue_button_test_kit/generated/v2.0.0/c4bb_test_suite'
 require_relative 'carin_for_blue_button_test_kit/generated/v2.0.0-dev-nonfinancial/c4bb_test_suite'
-require_relative 'carin_for_blue_button_test_kit/client/c4bb_client_test_suite'
+require_relative 'carin_for_blue_button_test_kit/client/v2.0.0/c4bb_client_test_suite'
