--- lib/carin_for_blue_button_test_kit/client/tags.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module CarinForBlueButtonTestKit
-  AUTH_TAG = 'carin_auth'
-  SUBMIT_TAG = 'carin_submit'
-  BINARY_TAG = 'carin_binary'
-  EXPORT_TAG = 'carin_export'
-  EXPORT_STATUS_TAG = 'carin_export_status'
-  EVERYTHING_TAG = 'carin_everything'
-  MEMBER_MATCH_TAG = 'carin_member_match'
-end
