+++ lib/carin_for_blue_button_test_kit/client/v2.0.0/claim_data_request_tests/initial_scratch_storing.rb
