+++ lib/carin_for_blue_button_test_kit/client/v2.0.0/endpoints/submit_claims_endpoint.rb
@@ -1,0 +1,24 @@
+require_relative '../tags'
+require_relative '../mock_server'
+
+module CarinForBlueButtonTestKit
+  class SubmitClaimsEndpoint < Inferno::DSL::SuiteEndpoint
+    include CarinForBlueButtonTestKit::MockServer
+
+    def test_run_identifier
+      extract_bearer_token(request)
+    end
+
+    def make_response
+      carin_resource_response(request)
+    end
+
+    def tags
+      [SUBMIT_TAG]
+    end
+
+    def update_result
+      results_repo.update(result.id, result: 'pass') unless test.config.options[:accepts_multiple_requests]
+    end
+  end
+end
