+++ lib/carin_for_blue_button_test_kit/client/v2.0.0/tags.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+module CarinForBlueButtonTestKit
+  AUTH_TAG = 'carin_auth'
+  SUBMIT_TAG = 'carin_submit'
+end
