+++ lib/carin_for_blue_button_test_kit/client/v2.0.0/user_input_response.rb
